requirements.txt
flake8
pytest
pdfplumber
nltk
black
ruff
autoflake
isort
pymongo
python-dotenv
scikit-learn
groq
fastapi
uvicorn[standard]
python-multipart
